+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandCreateNewExtractionConfigurationForProject.cs
@@ -4,26 +4,84 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using System;
+using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
+using Rdmp.Core.Curation.Data;
 using Rdmp.Core.DataExport.Data;
 using Rdmp.Core.Icons.IconProvision;
+using Rdmp.Core.Providers;
+using Rdmp.Core.Repositories.Construction;
 using ReusableLibraryCode.Icons.IconProvision;
 
 namespace Rdmp.Core.CommandExecution.AtomicCommands
 {
-    public class ExecuteCommandCreateNewExtractionConfigurationForProject:BasicCommandExecution,IAtomicCommand
+    /// <summary>
+    /// Creates a new <see cref="ExtractionConfiguration"/> under a given <see cref="Project"/>
+    /// </summary>
+    public class ExecuteCommandCreateNewExtractionConfigurationForProject : BasicCommandExecution, IAtomicCommand
     {
         private readonly Project _project;
+        private readonly string _name;
+        private ExtractableCohort cohort;
 
-        public ExecuteCommandCreateNewExtractionConfigurationForProject(IBasicActivateItems activator,Project project) : base(activator)
+        /// <summary>
+        /// True to prompt the user to pick an <see cref="ExtractableCohort"/> after creating the <see cref="ExtractionConfiguration"/>
+        /// </summary>
+        public bool PromptForCohort { get; set; } = true;
+
+        /// <summary>
+        /// True to prompt the user to pick some <see cref="ExtractableDataSet"/> after creating the <see cref="ExtractionConfiguration"/>
+        /// </summary>
+        public bool PromptForDatasets { get; set; } = true;
+
+        /// <summary>
+        /// An explicit cohort to assign for the <see cref="ExtractionConfiguration"/>
+        /// </summary>
+        public ExtractableCohort CohortIfAny { get => cohort; set {
+
+                if (!GetProjects(value).Any())
+                {
+                    SetImpossible($"There are no Projects with ProjectNumber {value.ExternalProjectNumber}");
+                }
+                cohort = value; } 
+        }
+
+        private IEnumerable<Project> GetProjects(ExtractableCohort cohortIfAny)
+        {
+            if(cohortIfAny is null)
+            {
+                // no cohort so all Project are valid
+                return BasicActivator.RepositoryLocator.DataExportRepository.GetAllObjects<Project>();
+            }
+
+            // we have a cohort so can only create an ExtractionConfiguration for Projects that share
+            // the cohorts project number
+
+            if (BasicActivator.CoreChildProvider is DataExportChildProvider dx)
+            {
+                return dx.Projects.Where(p => p.ProjectNumber == cohortIfAny.ExternalProjectNumber);
+            }
+
+            return Enumerable.Empty<Project>();
+        }
+
+        [UseWithObjectConstructor]
+        public ExecuteCommandCreateNewExtractionConfigurationForProject(IBasicActivateItems activator,
+
+            [DemandsInitialization("The Project under which to create the new ExtractionConfiguration")]
+            Project project,
+            [DemandsInitialization("The name for the new ExtractionConfiguration")]
+            string name = "") : base(activator)
         {
             _project = project;
+            this._name = name;
         }
-        
+
         public ExecuteCommandCreateNewExtractionConfigurationForProject(IBasicActivateItems activator) : base(activator)
         {
-            if(!activator.GetAll<IProject>().Any())
+            if (!activator.GetAll<IProject>().Any())
                 SetImpossible("You do not have any projects yet");
         }
 
@@ -41,16 +99,76 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
         {
             base.Execute();
 
-            var p = _project ?? SelectOne<Project>(BasicActivator.RepositoryLocator.DataExportRepository);
+            var p = _project;
 
             if(p == null)
+            {
+                if (!SelectOne(new DialogArgs
+                {
+                    WindowTitle = "Select Project",
+                    TaskDescription = GetTaskDescription(),
+                }, GetProjects(CohortIfAny).ToList(), out p))
+                    return;
+
+            }
+            if (p == null)
                 return;
 
-            var newConfig = new ExtractionConfiguration(BasicActivator.RepositoryLocator.DataExportRepository, p);
+            string name = _name;
+
+            // if we don't have a name and we are running in interactive mode
+            if (string.IsNullOrWhiteSpace(name) && BasicActivator.IsInteractive)
+            {
+                if (!BasicActivator.TypeText(new DialogArgs
+                {
+                    WindowTitle = "New Extraction Configuration",
+                    TaskDescription = "Enter a name for the new Extraction Configuration",
+                    EntryLabel = "Name",
+                }, 255, $"{p.ProjectNumber} {DateTime.Now.ToString("yyyy-MM-dd")} Extraction".Trim(), out name, false))
+                    return;
+            }
+
+            // create the new config
+            var newConfig = new ExtractionConfiguration(BasicActivator.RepositoryLocator.DataExportRepository, p, name);
+
+            if(CohortIfAny != null)
+            {
+                newConfig.Cohort_ID = CohortIfAny.ID;
+                newConfig.SaveToDatabase();
+            }
+            else
+            {
+                var chooseCohort = new ExecuteCommandChooseCohort(BasicActivator, newConfig);
+                if (PromptForCohort && BasicActivator.IsInteractive && !chooseCohort.IsImpossible)
+                {
+                    chooseCohort.Execute();
+
+                    // user cancelled picking a cohort
+                    if (newConfig.Cohort_ID == null)
+                        return;
+                }
+            }
+
+
+            var chooseDatasetsCommand = new ExecuteCommandAddDatasetsToConfiguration(BasicActivator, newConfig);
+
+            if (PromptForDatasets && BasicActivator.IsInteractive && !chooseDatasetsCommand.IsImpossible)
+            {
+                chooseDatasetsCommand.Execute();
+            }
 
             //refresh the project
             Publish(p);
             Activate(newConfig);
+            Emphasise(newConfig);
+        }
+
+        private string GetTaskDescription()
+        {
+            if (CohortIfAny == null)
+                return "Select which Project to create the ExtractionConfiguration under";
+
+            return $"Select which Project to create the ExtractionConfiguration under.  Only Projects with ProjectNumber {CohortIfAny.ExternalProjectNumber} are shown.  This is because you are using ExtractableCohort '{CohortIfAny}' for this operation.";
         }
     }
 }