+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - "parameter description" and "property name" have been added to the "set value" option for filters - https://github.com/HicServices/RDMP/issues/1034
 - Filter parameter values are now prompted for the user when adding existing filter without known good value sets - https://github.com/HicServices/RDMP/issues/1030
 - "Set Parameter Value(s)" option added to filter menus so you can more easily change the parameter values - https://github.com/HicServices/RDMP/issues/1035
+- Add options to create an extraction from a Cohorts right click menu and main userinterface - https://github.com/HicServices/RDMP/issues/1039
+
 
 ### Fixed
 
@@ -22,7 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - New filters are now highlighted correctly when added to a CIC - https://github.com/HicServices/RDMP/issues/1031
+- Creating a new Extracion Configuration will now ask the user for Name, Cohort and Datasets to be included for the extraction - https://github.com/HicServices/RDMP/issues/983
 - AllowIdentifiableExtractions is now an ErrorCode so can be set to Success instead of always being Fail or Warning (i.e. to completley ignore it).
+
 
 ## [7.0.9] - 2022-03-29